Voice/requirements.txt
groq==0.11.0
python-dotenv==1.0.1
sounddevice==0.5.1
